Fixed #9104 -- Moved FieldDoesNotExist to core.exceptions
commit aedab36eaf9e7206ace894d050535168eb205f5b
Author: Daniel Pyrathon <pirosb3@gmail.com>
Date:   Fri Jan 2 16:10:33 2015 +0100

    added fixes for comments

commit 9ae2305c33c9949c563f1d6fd8965f5cb038846e
Author: Daniel Pyrathon <pirosb3@gmail.com>
Date:   Fri Jan 2 15:48:57 2015 +0100

    imported FieldDoesNotExist for backwards compatibility

commit 2040240feb0194be6e036d55413a8440d2f017b7
Author: Daniel Pyrathon <pirosb3@gmail.com>
Date:   Fri Jan 2 15:39:44 2015 +0100

    moved FieldDoesNotExist to core.exceptions
