+++ django/contrib/admin/validation.py
@@ -1,6 +1,5 @@
-from django.core.exceptions import ImproperlyConfigured
+from django.core.exceptions import FieldDoesNotExist, ImproperlyConfigured
 from django.db import models
-from django.db.models.fields import FieldDoesNotExist
 from django.forms.models import BaseModelForm, BaseModelFormSet, _get_foreign_key
 from django.contrib.admin.utils import get_fields_from_path, NotRelationField
 
@@ -42,7 +41,7 @@ class BaseValidator(object):
                     continue
                 try:
                     f = model._meta.get_field(field)
-                except models.FieldDoesNotExist:
+                except FieldDoesNotExist:
                     # If we can't find a field on the model that matches, it could be an
                     # extra field on the form; nothing to check so move on to the next field.
                     continue
@@ -196,7 +195,7 @@ class BaseValidator(object):
                         if not hasattr(model, field):
                             try:
                                 model._meta.get_field(field)
-                            except models.FieldDoesNotExist:
+                            except FieldDoesNotExist:
                                 raise ImproperlyConfigured(
                                     "%s.readonly_fields[%d], %r is not a callable or "
                                     "an attribute of %r or found in the model %r."
@@ -250,7 +249,7 @@ class ModelAdminValidator(BaseValidator):
                         if not hasattr(model, field):
                             try:
                                 model._meta.get_field(field)
-                            except models.FieldDoesNotExist:
+                            except FieldDoesNotExist:
                                 raise ImproperlyConfigured(
                                     "%s.list_display[%d], %r is not a callable or "
                                     "an attribute of %r or found in the model %r."
@@ -348,7 +347,7 @@ class ModelAdminValidator(BaseValidator):
             for idx, field_name in enumerate(cls.list_editable):
                 try:
                     field = model._meta.get_field_by_name(field_name)[0]
-                except models.FieldDoesNotExist:
+                except FieldDoesNotExist:
                     raise ImproperlyConfigured("'%s.list_editable[%d]' refers to a "
                         "field, '%s', not defined on %s.%s."
                         % (cls.__name__, idx, field_name, model._meta.app_label, model.__name__))
@@ -429,7 +428,7 @@ def check_isdict(cls, label, obj):
 def get_field(cls, model, label, field):
     try:
         return model._meta.get_field(field)
-    except models.FieldDoesNotExist:
+    except FieldDoesNotExist:
         raise ImproperlyConfigured("'%s.%s' refers to field '%s' that is missing from model '%s.%s'."
                 % (cls.__name__, label, field, model._meta.app_label, model.__name__))
 
@@ -437,7 +436,7 @@ def get_field(cls, model, label, field):
 def fetch_attr(cls, model, label, field):
     try:
         return model._meta.get_field(field)
-    except models.FieldDoesNotExist:
+    except FieldDoesNotExist:
         pass
     try:
         return getattr(model, field)
