+++ django/contrib/gis/sitemaps/views.py
@@ -4,8 +4,8 @@ from django.apps import apps
 from django.http import Http404
 from django.contrib.gis.db.models.fields import GeometryField
 from django.contrib.gis.shortcuts import render_to_kml, render_to_kmz
+from django.core.exceptions import FieldDoesNotExist
 from django.db import connections, DEFAULT_DB_ALIAS
-from django.db.models.fields import FieldDoesNotExist
 
 
 def kml(request, label, model, field_name=None, compress=False, using=DEFAULT_DB_ALIAS):
