+++ django/db/models/manager.py
@@ -2,9 +2,9 @@ import copy
 from importlib import import_module
 import inspect
 
+from django.core.exceptions import FieldDoesNotExist
 from django.db import router
 from django.db.models.query import QuerySet
-from django.db.models.fields import FieldDoesNotExist
 from django.utils import six
 from django.utils.encoding import python_2_unicode_compatible
