+++ django/db/models/query_utils.py
@@ -10,6 +10,7 @@ from __future__ import unicode_literals
 from collections import namedtuple
 
 from django.apps import apps
+from django.core.exceptions import FieldDoesNotExist
 from django.db.backends import utils
 from django.db.models.constants import LOOKUP_SEP
 from django.utils import six
@@ -99,7 +100,6 @@ class DeferredAttribute(object):
         Retrieves and caches the value from the datastore on the first lookup.
         Returns the cached value.
         """
-        from django.db.models.fields import FieldDoesNotExist
         non_deferred_model = instance._meta.proxy_for_model
         opts = non_deferred_model._meta
